Convert ImageJ 1.x plugins to SciJava commands
Among other benefits, this allows the plugins to be discovered
efficiently simply by being on the classpath, rather than needing a
plugins.dir system property pointing to a single location with all
the JAR files, which must then be scanned for plugins.config files.

This is a breaking API change, so bumps to major version 4.